fix: お店タブの dnd-kit を遅延読み込み (#145 Phase 4)
- SortableRestaurantTabs に DndContext / SortableContext を集約
- RestaurantTabsStatic を初回フォールバック（ドラッグ不可）
- RestaurantTabsLazy で dynamic import、TodayClient は arrayMove と DragEndEvent 型のみ sortable/core に依存

Issue #145 Phase 4

Made-with: Cursor
